src/requirements.txt
requests
Pillow
simplejson
